Installation tests now use the actual installer
